% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasyshape.R
\name{reasyshape}
\alias{reasyshape}
\title{reasyshape}
\usage{
reasyshape(...)
}
\arguments{
\item{...}{Passed onto \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\description{
Run the \code{reasyshape} interactive shiny app.
}
\examples{
\dontrun{
library(reasyshape)
reasyshape()
}
}
